Field: Compatibility with HL7 spec numbering
Add an empty element in position 0 to index from 1. From now on, the Y in msg[N][Y][Y][Y][Y] are using a numbering that's compliant with HL7. The last N left (index of the segment in the message) is not commonly used in HL7, it will thus be left as a 0-indexed number.